refactor: move to utils in eth folder
