+++ src/index.ts
@@ -9,4 +9,3 @@ export * from './KeyringSnapRpcClient';
 export * from './rpc-handler';
 export * from './snap-utils';
 export * from './superstruct';
-export * from './helpers';
